requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dateutil>=2.9.0
PyYAML>=6.0.1
pdfplumber>=0.11.0
